+++ web/app/connectors/_merlins-connector/account/commands.js
@@ -38,7 +38,7 @@ import {
     parseOrder,
     parseAccountLocations
 } from './parsers'
-import {jqueryAjaxWrapper} from '../utils'
+import {jqueryAjaxWrapper, updateLoggedInState} from '../utils'
 import {
     CART_URL,
     LOGIN_POST_URL,
@@ -48,8 +48,6 @@ import {
     WISHLIST_URL,
     getWishlistQuantityUrl
 } from '../config'
-
-import {setLoggedIn} from 'progressive-web-sdk/dist/integration-manager/results'
 
 export const initLoginPage = (url) => (dispatch) => {
     return dispatch(fetchPageData(url))
@@ -146,7 +144,7 @@ const clearMessageCookie = () => {
 const DEFAULT_ERROR_TEXT = 'Username or password is incorrect'
 const EXISTING_ACCT_REGEX = /already an account/
 
-const submitForm = (href, formValues, formSelector, responseUrl) => {
+const submitForm = (href, formValues, formSelector) => {
     clearMessageCookie()
     return makeFormEncodedRequest(href, formValues, {method: 'POST'})
         .then(jqueryResponse)
@@ -155,6 +153,7 @@ const submitForm = (href, formValues, formSelector, responseUrl) => {
         })
         .then((res) => {
             const [$, $response] = res // eslint-disable-line no-unused-vars
+
             if (isFormResponseInvalid($response, formSelector)) {
                 const messages = JSON.parse(decodeURIComponent(getCookieValue(MAGENTO_MESSAGE_COOKIE)))
 
@@ -172,7 +171,7 @@ const submitForm = (href, formValues, formSelector, responseUrl) => {
                     _error: message
                 })
             }
-            return responseUrl
+            return res
         })
 }
 
@@ -191,7 +190,9 @@ export const login = (username, password, rememberMe) => (dispatch, getState) =>
         formData.persistent_remember_me = 'on'
     }
 
-    return submitForm(LOGIN_POST_URL, formData, '.form-login', '/customer/account')
+    return submitForm(LOGIN_POST_URL, formData, '.form-login')
+        .then((res) => dispatch(updateLoggedInState(res)))
+        .then(() => '/customer/account')
 }
 
 export const registerUser = (firstname, lastname, email, password, rememberMe) => (dispatch, getState) => {
@@ -209,7 +210,8 @@ export const registerUser = (firstname, lastname, email, password, rememberMe) =
     if (rememberMe) {
         formData.persistent_remember_me = 'on'
     }
-    return submitForm(CREATE_ACCOUNT_POST_URL, formData, '.form-create-account', '/customer/account')
+    return submitForm(CREATE_ACCOUNT_POST_URL, formData, '.form-create-account')
+        .then(() => '/customer/account')
 }
 
 const findPathForRoute = (routes, routeName) => {
@@ -229,10 +231,11 @@ export const navigateToSection = (router, routes, sectionName) => {
 
 export const logout = () => (dispatch) => (
     makeRequest('/customer/account/logout/')
+        .then(jqueryResponse)
         // Don't wait for the cart to do everything else
-        .then(() => {
+        .then((res) => {
+            dispatch(updateLoggedInState(res))
             dispatch(getCart())
-            dispatch(setLoggedIn(false))
         })
         // Update navigation menu and logged in flag
         // Need to request current location so that the right entry is active
@@ -300,8 +303,9 @@ export const editAddress = (address, addressId) => (dispatch, getState) => { // 
         form_key: formKey,
         ...createAddressRequestObject(address)
     }
-    return submitForm(`/customer/address/formPost/id/${addressId}`, formData, '.form-address-edit', '/customer/address/index/')
+    return submitForm(`/customer/address/formPost/id/${addressId}`, formData, '.form-address-edit')
         .then(() => dispatch(updateCustomerAddresses()))
+        .then(() => '/customer/address/index/')
 }
 
 export const addAddress = (address) => (dispatch, getState) => {
@@ -310,8 +314,9 @@ export const addAddress = (address) => (dispatch, getState) => {
         form_key: formKey,
         ...createAddressRequestObject(address)
     }
-    return submitForm('/customer/address/formPost/', formData, '.form-address-edit', '/customer/address/index/')
+    return submitForm('/customer/address/formPost/', formData, '.form-address-edit')
         .then(() => dispatch(updateCustomerAddresses()))
+        .then(() => '/customer/address/index/')
 }
 
 /* eslint-disable camelcase */
@@ -330,7 +335,8 @@ export const updateAccountInfo = ({names, email, currentPassword, newPassword}) 
     }
 
     dispatch(receiveAccountInfoData({names, email}))
-    return submitForm('/customer/account/editPost/', formData, '.form-edit-account', '/customer/account/edit/')
+    return submitForm('/customer/account/editPost/', formData, '.form-edit-account')
+        .then(() => '/customer/account/edit/')
 }