+++ web/app/containers/header/actions.js
@@ -6,6 +6,14 @@
 import {createAction, createActionWithAnalytics} from 'progressive-web-sdk/dist/utils/action-creation'
 import {getSearchSuggestions, searchProducts} from 'progressive-web-sdk/dist/integration-manager/app/commands'
 import {EVENT_ACTION} from 'progressive-web-sdk/dist/analytics/data-objects/'
+import {browserHistory} from 'progressive-web-sdk/dist/routing'
+import {lockScroll, unlockScroll} from '../app/actions'
+
+export const pushHistoryItem = createAction('Added item to history stack')
+
+export const popHistoryItem = createAction('Removed item from history stack')
+
+export const setIsHistoryPage = createAction('Navigated to page from history', ['isHistoryPage'])
 
 const searchAnalytics = createActionWithAnalytics(
     'Send search analytics', [],
@@ -19,6 +27,16 @@ export const openSearch = createAction('Open header search')
 export const closeSearch = createAction('Close header search')
 export const clearSuggestions = createAction('Clear search suggestion')
 
+export const openSearchModal = () => (dispatch) => {
+    dispatch(openSearch())
+    dispatch(lockScroll())
+}
+
+export const closeSearchModal = () => (dispatch) => {
+    dispatch(closeSearch())
+    dispatch(unlockScroll())
+}
+
 export const searchQueryChanged = (query) => (dispatch) => (
     dispatch(getSearchSuggestions(query))
 )
@@ -26,4 +44,10 @@ export const searchQueryChanged = (query) => (dispatch) => (
 export const searchSubmit = (query) => (dispatch) => {
     dispatch(searchAnalytics(query))
     dispatch(searchProducts(query))
+}
+
+export const goBack = () => (dispatch) => {
+    dispatch(popHistoryItem())
+    dispatch(setIsHistoryPage(true))
+    return browserHistory.goBack()
 }